added death popup | refactored dialogfactory to other project
