+++ devDungeon/src/level/devlevel/riddleHandler/BridgeGuardRiddleHandler.java
@@ -4,6 +4,7 @@ import components.MagicShieldComponent;
 import contrib.components.HealthComponent;
 import contrib.components.InteractionComponent;
 import contrib.components.InventoryComponent;
+import contrib.entities.DialogFactory;
 import contrib.entities.IHealthObserver;
 import contrib.entities.MiscFactory;
 import contrib.hud.dialogs.OkDialog;
@@ -24,7 +25,6 @@ import core.level.utils.Coordinate;
 import core.level.utils.LevelElement;
 import core.utils.IVoidFunction;
 import core.utils.components.MissingComponentException;
-import entities.DialogFactory;
 import entities.levercommands.BridgeControlCommand;
 import item.concreteItem.ItemPotionAttackSpeedPotion;
 import java.util.ArrayList;
